requirements.txt
click>=8.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
Pillow>=10.0.0
imagehash>=4.3.1
httpx>=0.25.0
aiofiles>=23.2.1
numpy>=1.24.0
openai>=1.10.0
tenacity>=8.2.0
rich>=13.0.0
tqdm>=4.66.0
